requirements.txt
numpy
pandas
timm
transformers==4.46.2
SimpleITK
torchio
monai
tqdm
